update documentation for cookie customization
