--- src/components/CookiesSelection/CookiesSelection.mdx
@@ -1,69 +1,0 @@
-import { Canvas, Controls, Meta } from '@storybook/blocks';
-import * as CookiesSelectionStories from './CookiesSelection.stories';
-
-<Meta of={CookiesSelectionStories} />
-
-# CookiesSelection
-
-Le composant `CookiesSelection` est utilisé pour permettre à l'utilisateur de visualiser le cookies utilisés sur le site répartis en 3 catégories : 
- - Les cookies essentiels au fonctionnement du site (essentials)
- - Les cookies sur lesquels reposent des fonctionnalités optionnelles (functional)
- - Les cookies de mesure d'analyse d'audience (analytics)
-
- L'utilisateur peut choisir d'accepter ou de refuser les cookies de chaque catégorie à l’exception des cookies essentiels pour le fonctionnement du site.
-
-<Canvas of={CookiesSelectionStories.Default}  />
-
-# API
-
-<Controls of={CookiesSelectionStories.Default} />
-
-Si une catégorie n'est pas définie, elle n’apparaîtra pas dans le composant.
-Le composant émet un événement `submit` avec les catégories de cookies acceptés par l'utilisateur. Les cookies non essentiels doivent être chargés après l'événement `submit` pour respecter le RGPD.
-
-## Description du cookies avancée
-
-Pour mettre d'avantage que du text brut en description d'un cookie, il est possible de passer par un slot.
-Le nom du slot aura pour format `cookie-description-<cookie-name>`.
-
-Une propriété `cookie` est passée au slot contenant les informations du cookie.
-
-<Canvas
-	of={CookiesSelectionStories.CookieDescriptionSlot}
-	source={{
-		language: 'html',
-		format: 'dedent',
-		code: `
-		<script lang="ts" setup>
-			const items = {
-				"functional": [
-					{
-						name: 'contrast',
-						conservation: '1 an',
-					},
-					{
-						name: 'privacy',
-						conservation: '1 an',
-					},
-					{
-						name: 'cookie_policy',
-						description: 'Sauvegarde les préférences de cookies.',
-						conservation: '1 an',
-					},
-				]
-			};
-		</script>
-
-		<template>
-			<CookiesSelection :items="items" >
-				<template #cookie-description-contrast="{ cookie }">
-					voir : <a href="#">En savoir plus</a>
-				</template>
-				<template #cookie-description-privacy="{ cookie }">
-					<a href="#">Politique de confidentialité</a>
-				</template>
-			</CookiesSelection>
-		</template>
-		`
-	}}
-/>
